+++ pysui/sui/sui_txresults/complex_tx.py
@@ -20,6 +20,7 @@ from pysui.sui.sui_txresults.common import CoinRef, GenericOwnerRef, GenericRef,
 from pysui.sui.sui_types.collections import EventID
 
 
+# TODO: Deprecate
 @dataclass
 class MoveCallTx(SuiTxReturnType, DataClassJsonMixin):
     """sui_moveCall transaction."""
@@ -31,6 +32,7 @@ class MoveCallTx(SuiTxReturnType, DataClassJsonMixin):
     type_arguments: list[Any] = field(metadata=config(letter_case=LetterCase.CAMEL), default_factory=list)
 
 
+# TODO: Deprecate
 @dataclass
 class PayTx(SuiTxReturnType, DataClassJsonMixin):
     """sui_pay transaction."""
@@ -40,6 +42,7 @@ class PayTx(SuiTxReturnType, DataClassJsonMixin):
     recipients: list[str]
 
 
+# TODO: Deprecate
 @dataclass
 class PayAllSuiTx(SuiTxReturnType, DataClassJsonMixin):
     """sui_payAllSui transaction."""
@@ -48,6 +51,7 @@ class PayAllSuiTx(SuiTxReturnType, DataClassJsonMixin):
     recipient: str
 
 
+# TODO: Deprecate
 @dataclass
 class PaySuiTx(SuiTxReturnType, DataClassJsonMixin):
     """sui_paySui transaction."""
@@ -57,6 +61,7 @@ class PaySuiTx(SuiTxReturnType, DataClassJsonMixin):
     recipients: list[str]
 
 
+# TODO: Deprecate
 @dataclass
 class PublishTx(SuiTxReturnType, DataClassJsonMixin):
     """sui_publish transaction."""
@@ -64,6 +69,7 @@ class PublishTx(SuiTxReturnType, DataClassJsonMixin):
     disassembled: dict[str, str]
 
 
+# TODO: Deprecate
 @dataclass
 class TransferObjectTx(SuiTxReturnType, DataClassJsonMixin):
     """sui_transferObject transaction."""
@@ -72,6 +78,7 @@ class TransferObjectTx(SuiTxReturnType, DataClassJsonMixin):
     recipient: str
 
 
+# TODO: Deprecate
 @dataclass
 class TransferSuiTx(SuiTxReturnType, DataClassJsonMixin):
     """sui_transferSui transaction."""
@@ -80,6 +87,7 @@ class TransferSuiTx(SuiTxReturnType, DataClassJsonMixin):
     amount: int
 
 
+# TODO: Deprecate
 @dataclass
 class ChangeEpocTx(SuiTxReturnType, DataClassJsonMixin):
     """From sui_executeTransaction forms."""
@@ -89,6 +97,7 @@ class ChangeEpocTx(SuiTxReturnType, DataClassJsonMixin):
     storage_charge: int
 
 
+# TODO: Deprecate
 @dataclass
 class GenesisTx(SuiTxReturnType, DataClassJsonMixin):
     """From sui_executeTransaction forms."""
@@ -96,6 +105,7 @@ class GenesisTx(SuiTxReturnType, DataClassJsonMixin):
     objects: list[str]
 
 
+# TODO: Deprecate
 @dataclass
 class AuthSignerInfo(SuiTxReturnType, DataClassJsonMixin):
     """Authorized signer info."""
@@ -103,18 +113,6 @@ class AuthSignerInfo(SuiTxReturnType, DataClassJsonMixin):
     epoch: int
     signature: str
     signers_map: list[int]
-
-
-@dataclass
-class ProgrammableTransaction(SuiTxReturnType, DataClassJsonMixin):
-    """Programmable transaction type info."""
-
-    commands: list[dict]
-    inputs: list[dict]
-
-    # def __post_init__(self):
-    #     """."""
-    #     print("Here")
 
 
 _TRANSACTION_LOOKUP = {
@@ -127,7 +125,6 @@ _TRANSACTION_LOOKUP = {
     "TransferSui": TransferSuiTx,
     "ChangeEpoch": ChangeEpocTx,
     "Genesis": GenesisTx,
-    "ProgrammableTransaction": ProgrammableTransaction,
 }
 
 
@@ -142,24 +139,112 @@ class TxGasData(SuiTxReturnType, DataClassJsonMixin):
 
 
 @dataclass
+class ChangeEpoch(SuiTxReturnType, DataClassJsonMixin):
+    """A system transaction that will update epoch information on-chain."""
+
+    kind: str
+    computational_charge: int
+    epoch: int
+    epoch_start_timestamp_ms: int
+    storage_charge: int
+    storage_rebate: int
+
+
+@dataclass
+class Genesis(SuiTxReturnType, DataClassJsonMixin):
+    """A system transaction used for initializing the initial state of the chain."""
+
+    kind: str
+    objects: list[str]
+
+
+@dataclass
+class ConsensusCommitPrologue(SuiTxReturnType, DataClassJsonMixin):
+    """A system transaction marking the start of a series of transactions scheduled as part of a checkpoint."""
+
+    kind: str
+    commit_timestamp_ms: int
+    epoch: int
+    prologue_round: int = field(metadata=config(field_name="round"))
+
+
+@dataclass
+class ProgrammableTransaction(SuiTxReturnType, DataClassJsonMixin):
+    """A series of commands where the results of one command can be used in future commands."""
+
+    kind: str
+    commands: list[dict]
+    inputs: list[dict]
+
+
+@dataclass
 class TransactionData(SuiTxReturnType, DataClassJsonMixin):
     """Transaction Data."""
 
     sender: str
     message_version: str = field(metadata=config(letter_case=LetterCase.CAMEL))
     gas_data: TxGasData = field(metadata=config(letter_case=LetterCase.CAMEL))
-    transactions: list[dict] = field(default_factory=list)
+    transaction: dict = field(default_factory=dict)
 
     def __post_init__(self):
         """Post init processing.
 
-        Loops through the transactions list to resolve hydrating different transaction type classes
+        Resolves the transaction kind
         """
-        hydrated = []
-        for tx_dict in self.transactions:
-            key = list(tx_dict.keys())[0]
-            hydrated.append(_TRANSACTION_LOOKUP[key].from_dict(tx_dict[key]))
-        self.transactions = hydrated
+        if self.transaction:
+            tx_kind = self.transaction["kind"]
+            match tx_kind:
+                case "ChangeEpoch":
+                    self.transaction = ChangeEpoch.from_dict(self.transaction)
+                case "Genesis":
+                    self.transaction = Genesis.from_dict(self.transaction)
+                case "ProgrammableTransaction":
+                    self.transaction = ProgrammableTransaction.from_dict(self.transaction)
+                case "ConsensusCommitPrologue":
+                    self.transaction = ConsensusCommitPrologue.from_dict(self.transaction)
+                case _:
+                    raise ValueError(f"{tx_kind} unknown Transaction Kind")
+
+
+@dataclass
+class MultiSigPublicKey(SuiTxReturnType, DataClassJsonMixin):
+    """Multi-sig Public Key info if used to sign transactions."""
+
+    pk_map: dict[str, str]
+    threshold: int
+
+
+@dataclass
+class MultiSig(SuiTxReturnType, DataClassJsonMixin):
+    """Defined for multi-sig transaction signing."""
+
+    bitmap: str
+    multisig_pk: MultiSigPublicKey
+    sigs: list[dict[str, str]]
+
+    def __post_init__(self):
+        """."""
+
+
+@dataclass
+class Transaction(SuiTxReturnType, DataClassJsonMixin):
+    """Transaction encapsulates single and multi-sig signatures as well as specific transaction data."""
+
+    data: TransactionData
+    tx_signatures: list[Union[str, dict]] = field(metadata=config(letter_case=LetterCase.CAMEL))
+
+    def __post_init__(self):
+        """Post process to reconcile signatures."""
+        hydrate = []
+        for sig in self.tx_signatures:
+            if isinstance(sig, str):
+                hydrate.append(sig)
+            elif isinstance(sig, dict):
+                hydrate.append(MultiSig.from_dict(sig))
+            else:
+                print(f"No idea {sig}")
+                hydrate.append(sig)
+        self.tx_signatures = hydrate
 
 
 # Events
@@ -377,112 +462,12 @@ class Effects(SuiTxReturnType, DataClassJsonMixin):
         """
 
 
-dummy = {
-    "jsonrpc": "2.0",
-    "result": {
-        "digest": "6DWchPLVr1B6sfd5UkbzgyXS4kW7gMGvyjW9zjqEkCnF",
-        "transaction": {
-            "data": {
-                "messageVersion": "v1",
-                "transactions": [
-                    {
-                        "ProgrammableTransaction": {
-                            "inputs": [],
-                            "commands": [
-                                {
-                                    "Publish": {
-                                        "disassembled": {
-                                            "base": "// Move bytecode v6\nmodule 0.base {\nuse 0000000000000000000000000000000000000000000000000000000000000001::vector;\nuse 0000000000000000000000000000000000000000000000000000000000000002::dynamic_field;\nuse 0000000000000000000000000000000000000000000000000000000000000002::dynamic_object_field;\nuse 0000000000000000000000000000000000000000000000000000000000000002::object;\nuse 0000000000000000000000000000000000000000000000000000000000000002::transfer;\nuse 0000000000000000000000000000000000000000000000000000000000000002::tx_context;\n\n\nstruct Service has key {\n\tid: UID,\n\tadmin: address\n}\nstruct TrackerField has store, key {\n\tid: UID\n}\nstruct TrackerObjectField has store, key {\n\tid: UID\n}\nstruct ServiceTracker has key {\n\tid: UID,\n\tinitialized: bool,\n\thas_child_field: bool,\n\thas_child_object_field: bool,\n\tcount_accounts: u64\n}\nstruct Tracker has store, key {\n\tid: UID,\n\tinitialized: bool,\n\taccumulator: vector<u8>\n}\n\nis_owner(Arg0: &Service, Arg1: address): bool {\nB0:\n\t0: MoveLoc[0](Arg0: &Service)\n\t1: ImmBorrowField[0](Service.admin: address)\n\t2: ReadRef\n\t3: MoveLoc[1](Arg1: address)\n\t4: Eq\n\t5: Ret\n}\nadd_dynamic_field(Arg0: &mut ServiceTracker, Arg1: &mut TxContext) {\nB0:\n\t0: MoveLoc[1](Arg1: &mut TxContext)\n\t1: Call object::new(&mut TxContext): UID\n\t2: Pack[1](TrackerField)\n\t3: StLoc[2](loc0: TrackerField)\n\t4: CopyLoc[0](Arg0: &mut ServiceTracker)\n\t5: MutBorrowField[1](ServiceTracker.id: UID)\n\t6: LdConst[4](Vector(U8): [9, 100, 121, 110, 95, 102, 105, 101, 108, 100])\n\t7: MoveLoc[2](loc0: TrackerField)\n\t8: Call dynamic_field::add<vector<u8>, TrackerField>(&mut UID, vector<u8>, TrackerField)\n\t9: LdTrue\n\t10: MoveLoc[0](Arg0: &mut ServiceTracker)\n\t11: MutBorrowField[2](ServiceTracker.has_child_field: bool)\n\t12: WriteRef\n\t13: Ret\n}\nadd_dynamic_object_field(Arg0: &mut ServiceTracker, Arg1: &mut TxContext) {\nB0:\n\t0: MoveLoc[1](Arg1: &mut TxContext)\n\t1: Call object::new(&mut TxContext): UID\n\t2: Pack[2](TrackerObjectField)\n\t3: StLoc[2](loc0: TrackerObjectField)\n\t4: CopyLoc[0](Arg0: &mut ServiceTracker)\n\t5: MutBorrowField[1](ServiceTracker.id: UID)\n\t6: LdConst[5](Vector(U8): [13, 100, 121, 110, 95, 111, 98, 106, 95, 102, 105, 101, 108, 100])\n\t7: MoveLoc[2](loc0: TrackerObjectField)\n\t8: Call dynamic_object_field::add<vector<u8>, TrackerObjectField>(&mut UID, vector<u8>, TrackerObjectField)\n\t9: LdTrue\n\t10: MoveLoc[0](Arg0: &mut ServiceTracker)\n\t11: MutBorrowField[3](ServiceTracker.has_child_object_field: bool)\n\t12: WriteRef\n\t13: Ret\n}\nincrease_account(Arg0: &mut ServiceTracker) {\nB0:\n\t0: CopyLoc[0](Arg0: &mut ServiceTracker)\n\t1: ImmBorrowField[4](ServiceTracker.count_accounts: u64)\n\t2: ReadRef\n\t3: LdU64(1)\n\t4: Add\n\t5: MoveLoc[0](Arg0: &mut ServiceTracker)\n\t6: MutBorrowField[4](ServiceTracker.count_accounts: u64)\n\t7: WriteRef\n\t8: Ret\n}\npublic accounts_created(Arg0: &ServiceTracker): u64 {\nB0:\n\t0: MoveLoc[0](Arg0: &ServiceTracker)\n\t1: ImmBorrowField[4](ServiceTracker.count_accounts: u64)\n\t2: ReadRef\n\t3: Ret\n}\npublic transfer(Arg0: Tracker, Arg1: address) {\nB0:\n\t0: MoveLoc[0](Arg0: Tracker)\n\t1: MoveLoc[1](Arg1: address)\n\t2: Call transfer::transfer<Tracker>(Tracker, address)\n\t3: Ret\n}\npublic stored_count(Arg0: &Tracker): u64 {\nB0:\n\t0: MoveLoc[0](Arg0: &Tracker)\n\t1: ImmBorrowField[5](Tracker.accumulator: vector<u8>)\n\t2: VecLen(12)\n\t3: Ret\n}\npublic add_to_store(Arg0: &mut Tracker, Arg1: u8): u64 {\nB0:\n\t0: CopyLoc[0](Arg0: &mut Tracker)\n\t1: MutBorrowField[5](Tracker.accumulator: vector<u8>)\n\t2: MoveLoc[1](Arg1: u8)\n\t3: VecPushBack(12)\n\t4: MoveLoc[0](Arg0: &mut Tracker)\n\t5: FreezeRef\n\t6: Call stored_count(&Tracker): u64\n\t7: Ret\n}\npublic add_from(Arg0: &mut Tracker, Arg1: vector<u8>) {\nB0:\n\t0: MoveLoc[0](Arg0: &mut Tracker)\n\t1: MutBorrowField[5](Tracker.accumulator: vector<u8>)\n\t2: MoveLoc[1](Arg1: vector<u8>)\n\t3: Call vector::append<u8>(&mut vector<u8>, vector<u8>)\n\t4: Ret\n}\npublic has_value(Arg0: &Tracker, Arg1: u8): bool {\nB0:\n\t0: MoveLoc[0](Arg0: &Tracker)\n\t1: ImmBorrowField[5](Tracker.accumulator: vector<u8>)\n\t2: ImmBorrowLoc[1](Arg1: u8)\n\t3: Call vector::contains<u8>(&vector<u8>, &u8): bool\n\t4: Ret\n}\npublic drop_from_store(Arg0: &mut Tracker, Arg1: u8): u8 {\nB0:\n\t0: CopyLoc[0](Arg0: &mut Tracker)\n\t1: ImmBorrowField[5](Tracker.accumulator: vector<u8>)\n\t2: ImmBorrowLoc[1](Arg1: u8)\n\t3: Call vector::index_of<u8>(&vector<u8>, &u8): bool * u64\n\t4: StLoc[2](loc0: u64)\n\t5: BrFalse(7)\nB1:\n\t6: Branch(11)\nB2:\n\t7: MoveLoc[0](Arg0: &mut Tracker)\n\t8: Pop\n\t9: LdConst[2](U64: [2, 0, 0, 0, 0, 0, 0, 0])\n\t10: Abort\nB3:\n\t11: MoveLoc[0](Arg0: &mut Tracker)\n\t12: MutBorrowField[5](Tracker.accumulator: vector<u8>)\n\t13: MoveLoc[2](loc0: u64)\n\t14: Call vector::remove<u8>(&mut vector<u8>, u64): u8\n\t15: Ret\n}\npublic create_service(Arg0: &mut TxContext) {\nB0:\n\t0: CopyLoc[0](Arg0: &mut TxContext)\n\t1: FreezeRef\n\t2: Call tx_context::sender(&TxContext): address\n\t3: StLoc[1](loc0: address)\n\t4: MoveLoc[0](Arg0: &mut TxContext)\n\t5: Call object::new(&mut TxContext): UID\n\t6: MoveLoc[1](loc0: address)\n\t7: Pack[0](Service)\n\t8: Call transfer::freeze_object<Service>(Service)\n\t9: Ret\n}\npublic create_service_tracker(Arg0: &mut TxContext) {\nB0:\n\t0: CopyLoc[0](Arg0: &mut TxContext)\n\t1: FreezeRef\n\t2: Call tx_context::sender(&TxContext): address\n\t3: StLoc[1](loc0: address)\n\t4: MoveLoc[0](Arg0: &mut TxContext)\n\t5: Call object::new(&mut TxContext): UID\n\t6: LdTrue\n\t7: LdFalse\n\t8: LdFalse\n\t9: LdU64(0)\n\t10: Pack[3](ServiceTracker)\n\t11: MoveLoc[1](loc0: address)\n\t12: Call transfer::transfer<ServiceTracker>(ServiceTracker, address)\n\t13: Ret\n}\ninit(Arg0: &mut TxContext) {\nB0:\n\t0: CopyLoc[0](Arg0: &mut TxContext)\n\t1: Call create_service(&mut TxContext)\n\t2: MoveLoc[0](Arg0: &mut TxContext)\n\t3: Call create_service_tracker(&mut TxContext)\n\t4: Ret\n}\nentry public set_dynamic_field(Arg0: &mut ServiceTracker, Arg1: &mut TxContext) {\nB0:\n\t0: MoveLoc[0](Arg0: &mut ServiceTracker)\n\t1: MoveLoc[1](Arg1: &mut TxContext)\n\t2: Call add_dynamic_field(&mut ServiceTracker, &mut TxContext)\n\t3: Ret\n}\nentry public set_dynamic_object_field(Arg0: &mut ServiceTracker, Arg1: &mut TxContext) {\nB0:\n\t0: MoveLoc[0](Arg0: &mut ServiceTracker)\n\t1: MoveLoc[1](Arg1: &mut TxContext)\n\t2: Call add_dynamic_object_field(&mut ServiceTracker, &mut TxContext)\n\t3: Ret\n}\nentry public create_account(Arg0: &Service, Arg1: &mut ServiceTracker, Arg2: address, Arg3: &mut TxContext) {\nB0:\n\t0: CopyLoc[3](Arg3: &mut TxContext)\n\t1: FreezeRef\n\t2: Call tx_context::sender(&TxContext): address\n\t3: StLoc[4](loc0: address)\n\t4: MoveLoc[0](Arg0: &Service)\n\t5: MoveLoc[4](loc0: address)\n\t6: Call is_owner(&Service, address): bool\n\t7: BrFalse(9)\nB1:\n\t8: Branch(15)\nB2:\n\t9: MoveLoc[1](Arg1: &mut ServiceTracker)\n\t10: Pop\n\t11: MoveLoc[3](Arg3: &mut TxContext)\n\t12: Pop\n\t13: LdConst[0](U64: [0, 0, 0, 0, 0, 0, 0, 0])\n\t14: Abort\nB3:\n\t15: MoveLoc[1](Arg1: &mut ServiceTracker)\n\t16: Call increase_account(&mut ServiceTracker)\n\t17: MoveLoc[3](Arg3: &mut TxContext)\n\t18: Call object::new(&mut TxContext): UID\n\t19: LdTrue\n\t20: LdConst[6](Vector(U8): [0])\n\t21: Pack[4](Tracker)\n\t22: MoveLoc[2](Arg2: address)\n\t23: Call transfer::transfer<Tracker>(Tracker, address)\n\t24: Ret\n}\nentry public add_value(Arg0: &mut Tracker, Arg1: u8, Arg2: &mut TxContext) {\nB0:\n\t0: MoveLoc[0](Arg0: &mut Tracker)\n\t1: MoveLoc[1](Arg1: u8)\n\t2: Call add_to_store(&mut Tracker, u8): u64\n\t3: Pop\n\t4: Ret\n}\nentry public remove_value(Arg0: &mut Tracker, Arg1: u8, Arg2: &mut TxContext) {\nB0:\n\t0: MoveLoc[0](Arg0: &mut Tracker)\n\t1: CopyLoc[1](Arg1: u8)\n\t2: Call drop_from_store(&mut Tracker, u8): u8\n\t3: MoveLoc[1](Arg1: u8)\n\t4: Eq\n\t5: BrFalse(7)\nB1:\n\t6: Branch(9)\nB2:\n\t7: LdConst[3](U64: [3, 0, 0, 0, 0, 0, 0, 0])\n\t8: Abort\nB3:\n\t9: Ret\n}\nentry public add_values(Arg0: &mut Tracker, Arg1: vector<u8>, Arg2: &mut TxContext) {\nB0:\n\t0: MoveLoc[0](Arg0: &mut Tracker)\n\t1: MoveLoc[1](Arg1: vector<u8>)\n\t2: Call add_from(&mut Tracker, vector<u8>)\n\t3: Ret\n}\n}"
-                                        }
-                                    }
-                                }
-                            ],
-                        }
-                    }
-                ],
-                "sender": "0xabfb3da1ccf95f2c0976b46c421fb89a765913089186d6ef9cc275258460e885",
-                "gasData": {
-                    "payment": [
-                        {
-                            "objectId": "0x21de82323c47d92d787a640592dceb8d898989c40335d1341d466bde6e930bd4",
-                            "version": 1,
-                            "digest": "EfEbK299x7mXsqkSNDVYB5VRw9GcawLSYeTzEBU6MCQX",
-                        }
-                    ],
-                    "owner": "0xabfb3da1ccf95f2c0976b46c421fb89a765913089186d6ef9cc275258460e885",
-                    "price": 1,
-                    "budget": 3000,
-                },
-            },
-            "txSignatures": [
-                "AJGHQsxah45dBGnONFw8221MI3lSL5MdUmltigKlcE5i+N4eJJ5myjovgmySNpnQf9hgt1WJqtYOlIdXiKVqwQEtc4qqvxVrdF+oUUJnATaD25ZViWzx35HlqqsMMHgx3g=="
-            ],
-        },
-        "effects": {
-            "messageVersion": "v1",
-            "status": {"status": "success"},
-            "executedEpoch": 0,
-            "gasUsed": {"computationCost": 390, "storageCost": 234, "storageRebate": 0},
-            "transactionDigest": "6DWchPLVr1B6sfd5UkbzgyXS4kW7gMGvyjW9zjqEkCnF",
-            "created": [
-                {
-                    "owner": "Immutable",
-                    "reference": {
-                        "objectId": "0x764ae5e43c4f9170436ef8c675b70f2b045428ff48757082ba186d5964cf3479",
-                        "version": 1,
-                        "digest": "95AeE8mTRkPYMHwqWvDKiKGLcu1ziWtMHa4RBntLEUPp",
-                    },
-                },
-                {
-                    "owner": "Immutable",
-                    "reference": {
-                        "objectId": "0x86257ab5c44d8646833c3c8f0cf77fcf92c031056858357d1d1ad68439298f1f",
-                        "version": 2,
-                        "digest": "BaUKa65jTwSwRBG6HoQeYB5RTFFKMRTMQhEkTNVea2kC",
-                    },
-                },
-                {
-                    "owner": {"AddressOwner": "0xabfb3da1ccf95f2c0976b46c421fb89a765913089186d6ef9cc275258460e885"},
-                    "reference": {
-                        "objectId": "0xddbacae8aad6192ad69b855dd066d719f6e6b82a9ef55682fe69c94cfc6a8197",
-                        "version": 2,
-                        "digest": "HYzPEJNwoPHbndzcYXVV8Tm7ewCmQiCKZxEeHgc8fWdP",
-                    },
-                },
-            ],
-            "mutated": [
-                {
-                    "owner": {"AddressOwner": "0xabfb3da1ccf95f2c0976b46c421fb89a765913089186d6ef9cc275258460e885"},
-                    "reference": {
-                        "objectId": "0x21de82323c47d92d787a640592dceb8d898989c40335d1341d466bde6e930bd4",
-                        "version": 2,
-                        "digest": "5TUq3qBuJrpLapGTAZeWmCi31SMbpXg6JUuEnEh3SQth",
-                    },
-                }
-            ],
-            "gasObject": {
-                "owner": {"AddressOwner": "0xabfb3da1ccf95f2c0976b46c421fb89a765913089186d6ef9cc275258460e885"},
-                "reference": {
-                    "objectId": "0x21de82323c47d92d787a640592dceb8d898989c40335d1341d466bde6e930bd4",
-                    "version": 2,
-                    "digest": "5TUq3qBuJrpLapGTAZeWmCi31SMbpXg6JUuEnEh3SQth",
-                },
-            },
-            "eventsDigest": "HiyNFQd83yKwcHQiNtUtF83Ai5njApAP7RCfM1EYmRj2",
-            "dependencies": ["4oQmWVqp7o3Cqqr4nhvVskbSdRPfwquFDH3Eo8WZmihX"],
-        },
-        "timestampMs": 1678365852173,
-        "checkpoint": 7926,
-    },
-    "id": 1,
-}
-
-
 @dataclass
 class TxResponse(SuiTxReturnType, DataClassJsonMixin):
     """Transaction Result."""
 
-    digest: Optional[str] = field(default_factory=str)
+    digest: str
+    # digest: Optional[str] = field(default_factory=str)
     transaction: Optional[dict] = field(default_factory=dict)
     errors: Optional[list[str]] = field(default_factory=list)
     effects: Optional[dict] = field(default_factory=dict)
@@ -500,13 +485,15 @@ class TxResponse(SuiTxReturnType, DataClassJsonMixin):
         """
         hydrated = []
         for ev_dict in self.events:
-            key = list(ev_dict.keys())[0]
-            hydrated.append({key: _EVENT_LOOKUP[key].from_dict(ev_dict[key])})
+            event_type = ev_dict["type"]
+            event_content = ev_dict["content"]
+            hydrated.append({event_type: _EVENT_LOOKUP[event_type].from_dict(event_content)})
         self.events = hydrated
         if self.effects:
             self.effects = Effects.from_dict(self.effects)
         if self.transaction:
-            self.transaction = TransactionData.from_dict(self.transaction["data"])
+            self.transaction = Transaction.from_dict(self.transaction)
+        print("foo")
 
     @property
     def succeeded(self) -> bool:
@@ -547,8 +534,9 @@ class DryRunTxResult(SuiTxReturnType, DataClassJsonMixin):
         """
         hydrated = []
         for ev_dict in self.events:
-            key = list(ev_dict.keys())[0]
-            hydrated.append({key: _EVENT_LOOKUP[key].from_dict(ev_dict[key])})
+            event_type = ev_dict["type"]
+            event_content = ev_dict["content"]
+            hydrated.append({event_type: _EVENT_LOOKUP[event_type].from_dict(event_content)})
         self.events = hydrated
 
 
@@ -567,8 +555,12 @@ class TxInspectionResult(SuiTxReturnType, DataClassJsonMixin):
         """
         hydrated = []
         for ev_dict in self.events:
-            key = list(ev_dict.keys())[0]
-            hydrated.append({key: _EVENT_LOOKUP[key].from_dict(ev_dict[key])})
+            event_type = ev_dict["type"]
+            event_content = ev_dict["content"]
+            hydrated.append({event_type: _EVENT_LOOKUP[event_type].from_dict(event_content)})
+        # for ev_dict in self.events:
+        #     key = list(ev_dict.keys())[0]
+        #     hydrated.append({key: _EVENT_LOOKUP[key].from_dict(ev_dict[key])})
         self.events = hydrated
 
     @classmethod